Project update.

+++ KindPaws.Backend/src/KindPaws.Application/Volunteers/VolunteerHandlers/Delete/DeleteVolunteerHandler.cs
@@ -1,4 +1,5 @@
-﻿using KindPaws.Application.Volunteers.VolunteerHandlers.Delete.DTOs;
+﻿using FluentValidation;
+using KindPaws.Application.Validation;
 using KindPaws.Domain.Shared.Others;
 using KindPaws.Domain.Shared.ValueObjects.IDs;
 using Microsoft.Extensions.Logging;
@@ -9,27 +10,35 @@ public class DeleteVolunteerHandler
 {
     private readonly ILogger<DeleteVolunteerHandler> _logger;
     private readonly IVolunteersRepository _volunteersRepository;
+    private readonly IValidator<DeleteVolunteerCommand> _validator;
+
 
     public DeleteVolunteerHandler(
         IVolunteersRepository volunteersRepository,
-        ILogger<DeleteVolunteerHandler> logger)
+        ILogger<DeleteVolunteerHandler> logger,
+        IValidator<DeleteVolunteerCommand> validator)
     {
         _volunteersRepository = volunteersRepository;
         _logger = logger;
+        _validator = validator;
     }
 
-    public async Task<Result<Guid, Error>> HandleAsync(
-        DeleteVolunteerRequest request,
+    public async Task<Result<Guid, ErrorList>> HandleAsync(
+        DeleteVolunteerCommand command,
         CancellationToken cancellationToken = default)
     {
-        var volunteerId = VolunteerId.Create(request.VolunteerId).Value;
+        var validationResult = await _validator.ValidateAsync(command, cancellationToken);
+        if (!validationResult.IsValid)
+            return validationResult.ToErrorList();
+        
+        var volunteerId = VolunteerId.Create(command.VolunteerId).Value;
 
         var volunteerResult = await _volunteersRepository.GetByIdAsync(
             volunteerId,
             cancellationToken);
 
         if (volunteerResult.IsFailure)
-            return volunteerResult.Error;
+            return volunteerResult.Error.ToErrorList();
 
         var result = await _volunteersRepository.DeleteAsync(volunteerResult.Value, cancellationToken);
